+++ packages/styles/src/packages/box/index.ts
@@ -1,0 +1,19 @@
+import {
+  sprinkle,
+  backgroundColorProperties,
+  borderColorProperties,
+  borderStyleProperties,
+  spaceProperties,
+  cursorProperties,
+} from "./box.sprinkle.css";
+
+export const box = {
+  sprinkle,
+  properties: {
+    backgroundColor: backgroundColorProperties,
+    borderColor: borderColorProperties,
+    borderStyle: borderStyleProperties,
+    space: spaceProperties,
+    cursor: cursorProperties,
+  },
+};