+++ packages/styles/src/themes/mediaQueries.ts
@@ -1,0 +1,9 @@
+import tokens from "@nimbus-ds/tokens/dist/js/tokens";
+
+const { breakpoint } = tokens;
+
+export const mediaQueries = {
+  xs: () => `screen and (min-width: ${breakpoint.xs.value})`,
+  md: () => `screen and (min-width: ${breakpoint.md.value})`,
+  lg: () => `screen and (min-width: ${breakpoint.lg.value})`,
+};